// client/src/state/action-types/menuTypes.ts
export enum MenuTypes {
  GET_MENU_CATEGORIES_START = 'menu/get_menu_categories_start',
  GET_MENU_CATEGORIES_SUCCESS = 'menu/get_menu_categories_success',
  GET_MENU_CATEGORIES_ERROR = 'menu/get_menu_categories_error',
  GET_MENU_ITEMS_START = 'menu/get_menu_items_start',
  GET_MENU_ITEMS_SUCCESS = 'menu/get_menu_items_success',
  GET_MENU_ITEMS_ERROR = 'menu/get_menu_items_error',
  SET_SELECTED_MENU_ITEM = 'menu/set_selected_menu_item',
  START_UPDATING_MENU_ITEM_START = 'menu/start_updating_menu_item_start',
  START_UPDATING_MENU_ITEM_SUCCESS = 'menu/start_updating_menu_item_success',
  START_UPDATING_MENU_ITEM_ERROR = 'menu/start_updating_menu_item_error',
  UNSET_UPDATING_ITEM = 'menu/unset_updating_item',
  UNSET_SELECTED_ITEM = 'menu/unset_selected_item',
}